fix: update jwtService, remove jwtExtractor and implement proper feignClient

+++ src/main/java/epam/com/gym/crm/dto/request/trainer/TrainerWorkloadRequest.java
@@ -15,5 +15,5 @@ public class TrainerWorkloadRequest {
     private String lastName;
     private Boolean isActive;
     private Date trainingDate;
-    private Double trainingDuration;
+    private int trainingDuration;
 }
